refactor: RegisterGoods->ProductRegister로 이름 변경 및 constatnts 별도의 파일에 분리
