+++ src/constants/index.js
@@ -1,3 +1,4 @@
 export { BUTTON_SIZE } from './ButtonSize';
-export { SET_EXPIRATION, SET_SALES, SET_DELIVERY } from './constant';
-export { INITIAL_STATES } from './initialstates';
+export { SET_EXPIRATION, SET_SALES } from './salesPeriod';
+export { SET_DELIVERY } from './deliveryPeriod';
+export { INITIAL_STATES } from './initialStates';
